Refactor SushiWallet deployment script and update test configurations
- Removed the Forge formatting step from the GitHub Actions workflow.
- Cleaned up the SushiWallet deployment script for better readability and organization.
- Updated SushiWallet contract and interfaces to improve clarity and maintainability.
- Enhanced test scripts by refining event handling and ensuring consistent formatting.
- Ensured all files end with a newline for better compatibility with various tools.

+++ test/SushiWallet.s.t.sol
@@ -7,6 +7,7 @@ import {SushiWallet} from "../src/SushiWallet.sol";
 
 contract SushiWalletScriptTest is Test {
     SushiWalletScript public deployer;
+
     event WalletDeployed(address wallet);
 
     function setUp() public {
@@ -15,7 +16,7 @@ contract SushiWalletScriptTest is Test {
         } catch {
             console.log("Warning: No RPC URL provided, running in local environment");
         }
-        
+
         deployer = new SushiWalletScript();
     }
 
@@ -23,10 +24,10 @@ contract SushiWalletScriptTest is Test {
         // Generate a private key and corresponding address
         uint256 privateKey = 0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80;
         address deployerAddress = vm.addr(privateKey);
-        
+
         // Fund the deployer account
         vm.deal(deployerAddress, 1 ether);
-        
+
         // Set the private key in environment
         vm.setEnv("PRIVATE_KEY", vm.toString(privateKey));
 
@@ -36,10 +37,10 @@ contract SushiWalletScriptTest is Test {
 
         // Get the deployment logs
         Vm.Log[] memory logs = vm.getRecordedLogs();
-        
+
         // Get the deployed contract address from the WalletDeployed event
         address walletAddress = address(0);
-        for (uint i = 0; i < logs.length; i++) {
+        for (uint256 i = 0; i < logs.length; i++) {
             // Look for WalletDeployed event
             if (logs[i].topics[0] == keccak256("WalletDeployed(address)")) {
                 // The address is in the data field since it's not indexed
@@ -59,9 +60,9 @@ contract SushiWalletScriptTest is Test {
             size := extcodesize(walletAddress)
         }
         assertTrue(size > 0, "Contract not deployed");
-        
+
         // Check constructor parameters were set correctly
         assertEq(address(wallet.SUSHI_ROUTER()), deployer.SUSHI_ROUTER());
         assertEq(address(wallet.MASTER_CHEF()), deployer.MASTER_CHEF());
     }
-} 
+}
